infrastructure/docker/airflow/requirements.txt
apache-airflow-providers-snowflake==5.6.0
apache-airflow-providers-amazon==8.9.0
apache-airflow-providers-smtp==1.8.0
snowflake-connector-python==3.10.0
boto3==1.34.0
